[added] clothing loot [added] vibroblade 1-hand loot [fixed] crash when equipping instrument [fixed] bug that made players always not certified for their weapon [changed] pvp damage reduction from 66% to 75% [changed] woundsRatio on weapons is stored as a float instead of an int
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@233 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/tangible/weapons/Weapon.h
@@ -165,7 +165,7 @@ public:
 
 	int getIdealAccuracy();
 
-	int getWoundsRatio();
+	float getWoundsRatio();
 
 	int getArmorPiercing();
 
@@ -330,7 +330,7 @@ public:
 
 	int getIdealAccuracy();
 
-	int getWoundsRatio();
+	float getWoundsRatio();
 
 	int getArmorPiercing();
